refactor: 장르 매핑 함수 위치 변경
