+++ src/entities/concert/constants/filterMaps.ts
@@ -1,14 +1,4 @@
-import { GenreEnum } from "../../genre/types";
 import { StatusFilter } from "../types";
-
-const genreMap: Record<GenreEnum, string> = {
-  [GenreEnum.ALL]: "전체",
-  [GenreEnum.JPOP]: "J-POP",
-  [GenreEnum.ROCK_METAL]: "ROCK / METAL",
-  [GenreEnum.RAP_HIPHOP]: "RAP / HIPHOP",
-  [GenreEnum.INDIE]: "INDIE",
-  [GenreEnum.POP]: "POP",
-};
 
 const statusMap: Record<StatusFilter, string> = {
   [StatusFilter.ALL]: "전체",
@@ -18,4 +8,4 @@ const statusMap: Record<StatusFilter, string> = {
   [StatusFilter.CANCELED]: "공연취소",
 };
 
-export { genreMap, statusMap };
+export { statusMap };
